Change a `int' declaration into a `size_t'
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@9920 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/file.c
@@ -461,7 +461,7 @@ rt_public void file_append(FILE *f, FILE *other, EIF_INTEGER l)
 	/*	Append a copy of `otherile' to `f' */
 
 	static char buffer[BUFSIZ];
-	int amount;
+	size_t amount;
 
 	errno = 0;
     if (0 != fseek(other, 0, FS_START))
